requirements.txt
fastapi==0.135.3
uvicorn==0.40.0
langchain==1.2.9
langgraph==1.0.8
langchain-ollama==1.0.1
langchain-tavily==0.2.17
python-dotenv==1.2.1